+++ R/obic.R
@@ -4,6 +4,7 @@
 #' 
 #' @param dt (data.table) A data.table containing the data of the fields to calcualte the OBI
 #' @param add_relative_score (logical) Should the relative score be calculated? Defaults to TRUE
+#' @param add_recommendations (logical) Should recommendations be given to improve the OBI score? Defaults to TRUE
 #' 
 #' @details The argument add_relative_score calculates relative perforamce of the agricultural fields in the OBI. 
 #' To have a meaningfull relative score a large number of fields need be given as input in dt. 
@@ -11,7 +12,7 @@
 #' @import data.table
 #' 
 #' @export
-obic <- function(dt, add_relative_score = TRUE) {
+obic <- function(dt, add_relative_score = TRUE, add_recommendations = TRUE) {
   
   # Check inputs
   checkmate::assert_data_table(dt)
@@ -39,9 +40,22 @@ obic <- function(dt, add_relative_score = TRUE) {
   # Score the fields
   dt.score <- OBIC::obic_score(dt.ind, add_relative_score = add_relative_score)
   
-  # Run the managment measures
-  dt.recom <- OBIC::obic_recommendations(dt.score)
+  if (add_recommendations) {
+    
+    # evaluate measures
+    dt.measure <- OBIC::obic_evalmeasure(dt.score)
+    
+    # make recommendations of top 3 measures
+    dt.recom <- OBIC::obic_recommendations(dt.measure, extensive = FALSE)
+    
+    result <- dt.recom
+    
+  } else {
+    
+    result <- dt.score
+    
+  }
   
-  return(dt.recom)
+  return(result)
   
 }